Moves Ranges into circuit_specs.py
